nightly v1.8.0.0 (12/18)

+++ Assets/QuantumUser/Simulation/NSMB/Entity/Player/MarioPlayerSystem.cs
@@ -2075,7 +2075,7 @@ namespace Quantum {
             var filter = f.Filter<MarioPlayer>();
             filter.UseCulling = false;
             while (filter.NextUnsafe(out EntityRef entity, out MarioPlayer* mario)) {
-                mario->Lives = f.Global->Rules.Lives;
+                mario->Lives = (byte) f.Global->Rules.Lives;
                 mario->PreRespawn(f, entity, stage);
             }
         }
